# Documentacion/openapi.yaml
openapi: 3.0.3
info:
  title: Plants App API
  description: |
    API REST del TFG "Plants App" — aplicación móvil de gestión de plantas con
    funcionalidades de comunidad, chat en tiempo real, recordatorios y datos climáticos.
    Arquitectura: Clean Architecture (Node.js + Express + MongoDB + Socket.IO).
  version: 1.0.0
  contact:
    name: TFG — 2 DAM
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo local

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTES COMUNES
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido tras login o registro.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
          example: Token inválido o expirado.
      required: [status, code, message]

    # ── Paginación ────────────────────────────────────────────────────────────
...
    UserPublic:
      type: object
      properties:
        id:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d1
        name:
          type: string
          example: Ana García
        email:
          type: string
          example: ana@example.com
        photo:
          type: string
          nullable: true
          example: https://res.cloudinary.com/demo/image/upload/avatar.jpg
        bannerPhoto:
          type: string
          nullable: true
          example: https://res.cloudinary.com/demo/image/upload/banner.jpg
        bio:
          type: string
          nullable: true
          example: Amante de las plantas tropicales.
        location:
          type: string
          nullable: true
          example: Sevilla
        locationLat:
          type: number
          format: double
          nullable: true
          example: 37.3891
        locationLon:
          type: number
          format: double
          nullable: true
          example: -5.9845
        role:
          type: string
          enum: [user, admin]
          example: user
        bannedUntil:
          type: string
          format: date-time
          nullable: true
          description: Fecha hasta la que el usuario está baneado. null si no está baneado.
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
          example: '2026-01-15T10:00:00.000Z'
    UserPreferences:
      type: object
      properties:
        appearInChatSearch:
          type: boolean
          example: true
        considerWeatherByDefault:
          type: boolean
          example: false
        isPrivate:
          type: boolean
          default: false
          description: Si true, el perfil no aparece en el feed público ni acepta nuevas conversaciones.
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        bio:
          type: string
          maxLength: 200
        location:
          type: string
          description: Nombre de ciudad o texto libre. Para coords usar locationLat/locationLon.
        locationLat:
          type: number
          format: double
          description: Latitud de la ubicación seleccionada del catálogo de capitales.
        locationLon:
          type: number
          format: double
          description: Longitud de la ubicación seleccionada del catálogo de capitales.
        photo:
          type: string
          description: URL Cloudinary de la foto de perfil.
        bannerPhoto:
          type: string
          description: URL Cloudinary del banner/fondo de perfil.
    UpdatePreferencesRequest:
      type: object
      properties:
        appearInChatSearch:
          type: boolean
        considerWeatherByDefault:
          type: boolean
        isPrivate:
          type: boolean
    UpdatePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          example: OldPass1!
        newPassword:
          type: string
          minLength: 8
          example: NewPass2@

    # ── Plantas ───────────────────────────────────────────────────────────────
    Plant:
      type: object
      properties:
        id:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d2
        userId:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d1
        name:
          type: string
          example: Monstera Deliciosa
        speciesId:
          type: string
          nullable: true
          description: ID de la especie del catálogo.
          example: 64f1a2b3c4d5e6f7a8b9c0e1
        photo:
          type: string
          nullable: true
          example: https://res.cloudinary.com/demo/image/upload/plant.jpg
        location:
          type: string
          nullable: true
          description: Ubicación del usuario propietario.
        plantLocation:
          type: string
          nullable: true
          description: Ubicación específica de la planta (ciudad).
        plantLocationLat:
          type: number
          format: double
          nullable: true
        plantLocationLon:
          type: number
          format: double
          nullable: true
        wateringFrequency:
          type: integer
          description: Frecuencia de riego en días.
          example: 7
        lightNeed:
          type: string
          enum: [low, medium, high]
          nullable: true
          description: Necesidad de luz.
        pruningFrequency:
          type: integer
          nullable: true
          description: Frecuencia de poda en días.
        notes:
          type: string
          nullable: true
        nextWatering:
          type: string
          format: date-time
          nullable: true
          example: '2026-03-07T08:00:00.000Z'
        nextPruning:
          type: string
          format: date-time
          nullable: true
        considerWeatherForWatering:
          type: boolean
          description: Si true, el cron de clima ajusta el riego según previsión.
        pendingRainAdjustment:
          type: object
          nullable: true
          description: |
            Reset de `nextWatering` pendiente de validación por el cron del
            día siguiente (decisión #169 / F7-043). Se rellena cuando
            `_processWeather` aplica un reset por previsión de lluvia y se
            limpia / aplica rollback en `_processYesterdayRain`.
          properties:
            resetAt:
              type: string
              format: date-time
              description: Momento en el que se aplicó el reset por previsión.
            previousNextWatering:
              type: string
              format: date-time
              nullable: true
              description: Valor de `nextWatering` ANTES del reset (null si no había).
            expectedMm:
              type: number
              description: Probabilidad/precipitación esperada que motivó el reset.
            locationLat:
              type: number
            locationLon:
              type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: No autorizado — token ausente o inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Prohibido — rol insuficiente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Error de validación en el cuerpo de la petición.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /users/search:
    get:
      tags: [Users]
      summary: Buscar usuarios por nombre
      description: Busca usuarios activos por nombre con regex case-insensitive. Admin ve todos; usuario normal excluye isPrivate=true. Excluye al propio solicitante. Máximo 20 resultados.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Texto de búsqueda (nombre).
      responses:
        '200':
          description: Lista de usuarios coincidentes.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    photo:
                      type: string
                      nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me:
    get:
      tags: [Users]
      summary: Obtener perfil del usuario autenticado
      responses:
        '200':
          description: Perfil del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Users]
      summary: Actualizar perfil del usuario autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Perfil actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Users]
      summary: Eliminar cuenta del usuario autenticado
      description: |
        Soft-delete de la cuenta. Los datos del usuario se marcan como eliminados.
        Si `preserveContent=false` (default), los posts y comentarios del usuario también se soft-delete.
        Si `preserveContent=true`, los posts y comentarios permanecen pero quedan sin autor (anónimos).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  description: Contraseña actual del usuario para confirmar la eliminación.
                preserveContent:
                  type: boolean
                  default: false
                  description: Si true, los posts y comentarios permanecen anónimos tras borrar la cuenta.
      responses:
        '204':
          description: Cuenta eliminada correctamente.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/export:
    get:
      tags: [Users]
      summary: Exportar datos personales del usuario (RGPD)
      description: Devuelve un JSON con todos los datos personales del usuario autenticado — perfil y plantas. Conforme a RGPD Art. 20 (portabilidad de datos).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Datos exportados correctamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  exportedAt:
                    type: string
                    format: date-time
                    example: '2026-03-09T10:00:00.000Z'
                  profile:
                    $ref: '#/components/schemas/UserPublic'
                  plants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Plant'
                  totalPlants:
                    type: integer
                    example: 5
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me/preferences:
    put:
      tags: [Users]
      summary: Actualizar preferencias del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
      responses:
        '200':
          description: Preferencias actualizadas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/password:
    put:
      tags: [Users]
      summary: Cambiar contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '204':
          description: Contraseña cambiada correctamente.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/fcm-token:
    put:
      tags: [Users]
      summary: Registrar / actualizar el token FCM del dispositivo
      description: |
        Persiste el token FCM del dispositivo del usuario autenticado para
        recibir notificaciones push (decisión #170 / F7-045). El frontend lo
        envía tras el login (vía `FirebaseMessagingService.registerToken`) y
        cada vez que Firebase rote el token (`onTokenRefresh`).

        Pasar un string vacío equivale a desregistrar el dispositivo.

        **Desasociación cross-cuenta (decisión #195 / F7-097):** si el mismo
        `fcmToken` ya estaba asignado a otro usuario (transferencia del
        dispositivo entre cuentas), el backend limpia primero el campo
        `fcmToken` del usuario anterior antes de asignarlo al actual. De
        este modo las push pendientes no llegan al usuario equivocado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fcmToken]
              properties:
                fcmToken:
                  type: string
                  description: Token FCM del dispositivo (vacío para desregistrar).
                  example: 'dQw4w9WgXcQ:APA91bH…'
      responses:
        '204':
          description: Token registrado correctamente.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Users]
      summary: Borrar el token FCM del usuario autenticado
      description: |
        Desregistro explícito del `fcmToken` del usuario actual (decisión
        #195 / F7-097). Llamado por el `LogoutUseCase` profundo del
        frontend durante el cierre de sesión para que las push pendientes
        no lleguen al dispositivo asociado a la sesión anterior.

        Idempotente: si el campo ya estaba vacío, devuelve 204 igualmente.
      responses:
        '204':
          description: Token borrado correctamente.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID del usuario.
    get:
      tags: [Users]
      summary: Obtener perfil público de un usuario
      responses:
        '200':
          description: Perfil público del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════════════════════════════════
  # REMINDERS
  # ════════════════════════════════════════════════════════════════════════════
...
  /admin/users/search:
    get:
      tags: [Admin]
      summary: Buscar usuarios por nombre o email
      description: Busca usuarios activos (no eliminados) por nombre o email con regex case-insensitive. Máximo 20 resultados.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Texto de búsqueda (nombre o email).
      responses:
        '200':
          description: Lista de usuarios coincidentes.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    email:
                      type: string
                    photo:
                      type: string
                      nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...